Refactor GUI renderer directories
