+++ packages/gui/src/renderer/__tests__/llm-bridge-store.test.ts
@@ -1,6 +1,6 @@
 import { tmpdir } from 'node:os';
 import path from 'node:path';
-import { LlmBridgeStore, LlmBridgeConfig } from '../llm-bridge-store';
+import { LlmBridgeStore, LlmBridgeConfig } from '../stores/llm-bridge-store';
 
 const tempDir = path.join(tmpdir(), 'bridge-store-test');
