+++ packages/gui/src/renderer/app/ChatApp.tsx
@@ -1,22 +1,23 @@
 import React from 'react';
-import { BridgeManager } from './BridgeManager';
-import EchoBridge from './bridges/EchoBridge';
-import ReverseBridge from './bridges/ReverseBridge';
+import { BridgeManager } from '../utils/BridgeManager';
+import EchoBridge from '../bridges/EchoBridge';
+import ReverseBridge from '../bridges/ReverseBridge';
 import { ChatManager, ChatSessionDescription, McpConfig } from '@agentos/core';
-import { createChatManager } from './chat-manager';
-import ChatSidebar from './ChatSidebar';
-import ChatTabs from './ChatTabs';
-import McpSettings from './McpSettings';
-import McpList from './McpList';
-import SettingsMenu from './SettingsMenu';
-import { McpConfigStore } from './mcp-config-store';
-import { loadMcpFromStore } from './mcp-loader';
-import PresetSelector from './PresetSelector';
-import { Preset, loadPresets, PresetStore } from './preset-store';
-import { LlmBridgeStore } from './llm-bridge-store';
-import ChatMessageList from './ChatMessageList';
-import ChatInput from './ChatInput';
-import useChatSession from './useChatSession';
+import { createChatManager } from '../utils/chat-manager';
+import ChatSidebar from '../components/ChatSidebar';
+import ChatTabs from '../components/ChatTabs';
+import McpSettings from '../pages/McpSettings';
+import McpList from '../pages/McpList';
+import SettingsMenu from '../components/SettingsMenu';
+import { McpConfigStore } from '../stores/mcp-config-store';
+import { loadMcpFromStore } from '../utils/mcp-loader';
+import PresetSelector from '../components/PresetSelector';
+import { Preset } from '@agentos/core';
+import { loadPresets, PresetStore } from '../stores/preset-store';
+import { LlmBridgeStore } from '../stores/llm-bridge-store';
+import ChatMessageList from '../components/ChatMessageList';
+import ChatInput from '../components/ChatInput';
+import useChatSession from '../hooks/useChatSession';
 
 const manager = new BridgeManager();
 manager.register('echo', new EchoBridge());
