+++ packages/gui/src/renderer/app/index.html
