+++ packages/gui/src/renderer/components/ChatInput.tsx
