+++ packages/gui/src/renderer/components/ChatSidebar.tsx
