+++ packages/gui/src/renderer/components/ChatTabs.tsx
