+++ packages/gui/src/renderer/pages/McpList.tsx
