+++ packages/gui/src/renderer/pages/McpSettings.tsx
