+++ packages/gui/src/renderer/stores/mcp-config-store.ts
