+++ packages/gui/src/renderer/stores/preset-store.ts
