+++ packages/gui/src/renderer/utils/BridgeManager.ts
