+++ packages/gui/src/renderer/utils/NoopCompressor.ts
