+++ packages/gui/src/renderer/utils/chat-manager.ts
